Project_Fire/backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
twilio==8.10.0
aiohttp==3.9.0
python-dotenv==1.0.0
pydantic==2.4.2
 Pillow==10.0.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-jose[cryptography]
bcrypt==4.0.1
email-validator==2.1.0